MercsAndBots -- Changed merc stances, grades, and counts to variable arrays. This allows for stances to be listed properly on both the UI and in-game without having to hardcode the number of stances, merc grade, or number of total mercs
NOTE: Each encode will have to be set up to handle the variable size. The size of the packet prior to this encode should not matter except for logging purposes. An example of how to do so is listed in the SoD.cpp encode for MercenaryListEntry_Struct's opcode.

There's also a fix for a debug assert I was getting.