MercsAndBots - More work on mercs.
Trying to get persistent timers working (they are, to a point), as well as cross-zone mercenaries working. Suspend timer button is now functional as a result.

I think something is up with mercs not actually being removed from the entity list, as I can leave a zone and have the merc spawn just fine afterwards, but the moment I re-enter the other zone, it either a) crashes, or b) doesn't spawn my merc. Just notes.